created new views: Footer, ButtonBack, GameState and GameProgress; removed excess templates; changed functions
