--- js/templates/button-back.js
@@ -1,16 +1,0 @@
-import {getElementFromTemplate, header} from '../functions';
-
-const moduleButtonBack = () => {
-  const buttonBackTemplate = getElementFromTemplate(`
-    <div class="header__back">
-      <button class="back">
-        <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-        <img src="img/logo_small.svg" width="101" height="44">
-      </button>
-    </div>
-`);
-
-  header.appendChild(buttonBackTemplate);
-};
-
-export default moduleButtonBack;
