+++ js/templates/greeting.js
@@ -1,12 +1,12 @@
 import GreetingView from '../view/greeting-view';
-import {showScreen, header, removeGameProgressContainer} from '../functions';
+import {showScreen, renderHeader, removeGameProgressContainer} from '../functions';
 import rules from './rules';
 import {startNewGame} from "../data/game-data";
 
 const moduleGreeting = () => {
   const greeting = new GreetingView();
 
-  header.innerHTML = ``;
+  renderHeader();
   showScreen(greeting.element);
   removeGameProgressContainer();
   startNewGame();
